requirements.txt
Django>=5.2.2
Pillow>=10.0.0
django-anymail[brevo]>=10.0
pdfkit>=1.0.0
django-paypal>=2.0
razorpay>=2.0.0
weasyprint<66
requests>=2.28.0
pytz>=2023.3
social-auth-app-django>=5.2.0
cryptography>=41.0.0
qrcode>=7.4.2
pandas>=2.0.0
openpyxl>=3.1.0
emoji>=2.8.0
cloudinary>=1.36.0
django-q2>=1.6.0
redis>=7.0.1
channels>=4.0.0
daphne>=4.0.0